+++ lib/src/views/indicators/indicator_view.dart
