+++ lib/src/views/lab/hospital_view.dart
